Minor fixes to TracefileVisitor Cleanup in RTLogger - Remeber to remove debug files!

+++ ide/plugins/showtraceNextGen/src/main/java/org/overture/ide/plugins/showtraceNextGen/viewer/TraceData.java
@@ -174,7 +174,7 @@ public class TraceData
         
         if(!messages.containsKey(pmsgid))
         {
-        	messages.put(pmsgid, new tdMessage(pmsgid.intValue()));
+        	messages.put(pmsgid, new tdMessage(pmsgid));
         }
         
         return messages.get(pmsgid);
